+++ lib/regex-router.js
@@ -21,7 +21,7 @@ Router.prototype.default = function (handler) {
   this._default = handler
 }
 
-Router.prototype.match = function (req, res) {
+Router.prototype.match = function (req, res, next) {
   var url = req.url
   var len = this.routes.length
   var i = 0
@@ -31,12 +31,12 @@ Router.prototype.match = function (req, res) {
     output = route.regex.exec(url)
 
     if (output) {
-      route.handler(req, res, output)
+      route.handler(req, res, output, next)
       return
     }
   }
 
-  this._default(req, res)
+  this._default(req, res, next)
 }
 
 function noop (req, res) {
